requirements.txt
git+git://github.com/johnglover/modal
rdflib
scipy
scikits.audiolab
scikits.samplerate
h5py
matplotlib
pyFluidSynth
